// src/models/response/AuthLogin.tsx
// Generated by https://quicktype.io

export interface AuthLogin {
    acessoId:    number;
    login:       string;
    status:      string;
    tipoUsuario: string;
    token:       string;
    info:        Info;
}

export interface Info {
    idInstancia:          number;
    contratoId:           number;
    tipoUsuarioCliente:   TipoUsuarioCliente;
    nome:                 string;
    razaoSocial:          null;
    nomeFantasia:         null;
    cpf:                  string;
    cnpj:                 null;
    email:                string;
    contato1:             string;
    contato2:             null;
    sexo:                 null;
    statusUsuario:        null;
    dataCadastro:         number;
    dataAtualizacao:      null;
    cargo:                string;
    observacao:           null;
    endereco:             null;
    numero:               null;
    complemento:          null;
    bairro:               null;
    cidade:               null;
    uf:                   null;
    cep:                  null;
    clienteId:            number;
    exibeAbaValor:        string;
    colaboradorCargoId:   null;
    menuSistema:          MenuSistema[];
    collaboratorUser:     null;
    typeCollaboratorUser: null;
    charter:              number;
}

export interface MenuSistema {
    id:          number;
    descricao:   string;
    colaborador: string;
    usuario:     string;
}

export interface TipoUsuarioCliente {
    id:        number;
    descricao: string;
    status:    string;
    charter:   number;
}